+++ src/lib/components/ui/combobox/combobox-content.svelte
@@ -1,0 +1,42 @@
+<script lang="ts">
+	import { Combobox as ComboboxPrimitive } from "bits-ui";
+	import { cn, type WithoutChild } from "$lib/utils.js";
+	import type { ComponentProps } from "svelte";
+	import type { WithoutChildrenOrChild } from "$lib/utils.js";
+
+	import SelectPortal from "../select/select-portal.svelte";
+	import SelectScrollUpButton from "../select/select-scroll-up-button.svelte";
+	import SelectScrollDownButton from "../select/select-scroll-down-button.svelte";
+
+	let {
+		ref = $bindable(null),
+		class: className,
+		sideOffset = 4,
+		portalProps,
+		children,
+		...restProps
+	}: WithoutChild<ComboboxPrimitive.ContentProps> & {
+		portalProps?: WithoutChildrenOrChild<ComponentProps<typeof SelectPortal>>;
+	} = $props();
+</script>
+
+<SelectPortal {...portalProps}>
+	<ComboboxPrimitive.Content
+		bind:ref
+		{sideOffset}
+		data-slot="combobox-content"
+		class={cn(
+			"bg-popover text-popover-foreground data-[state=open]:animate-in data-[state=closed]:animate-out data-[state=closed]:fade-out-0 data-[state=open]:fade-in-0 data-[state=closed]:zoom-out-95 data-[state=open]:zoom-in-95 data-[side=bottom]:slide-in-from-top-2 data-[side=left]:slide-in-from-end-2 data-[side=right]:slide-in-from-start-2 data-[side=top]:slide-in-from-bottom-2 relative z-50 max-h-[300px] min-w-[8rem] origin-(--bits-select-content-transform-origin) overflow-x-hidden overflow-y-auto rounded-md border shadow-md data-[side=bottom]:translate-y-1 data-[side=left]:-translate-x-1 data-[side=right]:translate-x-1 data-[side=top]:-translate-y-1",
+			className
+		)}
+		{...restProps}
+	>
+		<SelectScrollUpButton />
+		<ComboboxPrimitive.Viewport
+			class={cn("w-full min-w-(--bits-combobox-anchor-width) scroll-my-1 p-1")}
+		>
+			{@render children?.()}
+		</ComboboxPrimitive.Viewport>
+		<SelectScrollDownButton />
+	</ComboboxPrimitive.Content>
+</SelectPortal>